Refactor: Transforms all epics into thunks
