+++ public/app/features/explore/state/actions.ts
@@ -14,12 +14,32 @@ import {
   hasNonEmptyQuery,
   buildQueryTransaction,
   updateHistory,
+  getRefIds,
+  instanceOfDataQueryError,
+  clearQueryKeys,
+  serializeStateToUrlParam,
 } from 'app/core/utils/explore';
 // Types
-import { ThunkResult } from 'app/types';
-import { DataSourceApi, DataQuery, DataSourceSelectItem, QueryFixAction, PanelData } from '@grafana/ui';
+import { ThunkResult, ExploreUrlState } from 'app/types';
+import {
+  DataSourceApi,
+  DataQuery,
+  DataSourceSelectItem,
+  QueryFixAction,
+  PanelData,
+  DataQueryResponseData,
+} from '@grafana/ui';
 
-import { RawTimeRange, LogsDedupStrategy, AbsoluteTimeRange, LoadingState } from '@grafana/data';
+import {
+  RawTimeRange,
+  LogsDedupStrategy,
+  AbsoluteTimeRange,
+  LoadingState,
+  DataFrame,
+  TimeRange,
+  isDateTime,
+  dateTimeForTimeZone,
+} from '@grafana/data';
 import { ExploreId, ExploreUIState, QueryTransaction, ExploreMode } from 'app/types/explore';
 import {
   updateDatasourceInstanceAction,
@@ -53,19 +73,24 @@ import {
   changeModeAction,
   scanStopAction,
   runQueriesAction,
-  stateSaveAction,
-  updateTimeRangeAction,
-  limitMessageRatePayloadAction,
   changeLoadingStateAction,
   historyUpdatedAction,
-  processQueryResultsAction,
-  processQueryErrorsAction,
   queryStartAction,
+  resetQueryErrorAction,
+  querySuccessAction,
+  queryFailureAction,
+  setUrlReplacedAction,
+  changeRangeAction,
 } from './actionTypes';
 import { ActionOf, ActionCreator } from 'app/core/redux/actionCreatorFactory';
 import { getTimeZone } from 'app/features/profile/state/selectors';
 import { offOption } from '@grafana/ui/src/components/RefreshPicker/RefreshPicker';
 import { getShiftedTimeRange } from 'app/core/utils/timePicker';
+import { ResultProcessor } from '../utils/ResultProcessor';
+import _ from 'lodash';
+import { toDataQueryError } from '../../dashboard/state/PanelQueryState';
+import { updateLocation } from '../../../core/actions';
+import { getTimeSrv } from '../../dashboard/services/TimeSrv';
 
 /**
  * Updates UI state and save it to the URL
@@ -73,7 +98,7 @@ import { getShiftedTimeRange } from 'app/core/utils/timePicker';
 const updateExploreUIState = (exploreId: ExploreId, uiStateFragment: Partial<ExploreUIState>): ThunkResult<void> => {
   return dispatch => {
     dispatch(updateUIStateAction({ exploreId, ...uiStateFragment }));
-    dispatch(stateSaveAction());
+    dispatch(stateSave());
   };
 };
 
@@ -171,7 +196,7 @@ export const updateTimeRange = (options: {
   absoluteRange?: AbsoluteTimeRange;
 }): ThunkResult<void> => {
   return dispatch => {
-    dispatch(updateTimeRangeAction({ ...options }));
+    dispatch(updateTime({ ...options }));
     dispatch(runQueries(options.exploreId));
   };
 };
@@ -193,7 +218,7 @@ export function clearQueries(exploreId: ExploreId): ThunkResult<void> {
   return dispatch => {
     dispatch(scanStopAction({ exploreId }));
     dispatch(clearQueriesAction({ exploreId }));
-    dispatch(stateSaveAction());
+    dispatch(stateSave());
   };
 }
 
@@ -256,7 +281,7 @@ export function initializeExplore(
         ui,
       })
     );
-    dispatch(updateTimeRangeAction({ exploreId }));
+    dispatch(updateTime({ exploreId }));
   };
 }
 
@@ -407,7 +432,7 @@ export function modifyQueries(
  */
 export function runQueries(exploreId: ExploreId): ThunkResult<void> {
   return (dispatch, getState) => {
-    dispatch(updateTimeRangeAction({ exploreId }));
+    dispatch(updateTime({ exploreId }));
 
     const exploreItemState = getState().explore[exploreId];
     const {
@@ -430,7 +455,7 @@ export function runQueries(exploreId: ExploreId): ThunkResult<void> {
 
     if (!hasNonEmptyQuery(queries)) {
       dispatch(clearQueriesAction({ exploreId }));
-      dispatch(stateSaveAction()); // Remember to save to state and update location
+      dispatch(stateSave()); // Remember to save to state and update location
       return;
     }
 
@@ -461,18 +486,12 @@ export function runQueries(exploreId: ExploreId): ThunkResult<void> {
       }
 
       if (state === LoadingState.Error) {
-        dispatch(processQueryErrorsAction({ exploreId, response: error, datasourceId }));
+        dispatch(processErrorResults({ exploreId, response: error, datasourceId }));
         return;
       }
 
       if (state === LoadingState.Streaming) {
-        dispatch(
-          limitMessageRatePayloadAction({
-            exploreId,
-            series: isLokiDataSource ? series : legacy,
-            datasourceId,
-          })
-        );
+        dispatch(limitMessageRate(exploreId, isLokiDataSource ? series : legacy, datasourceId));
         return;
       }
 
@@ -488,7 +507,7 @@ export function runQueries(exploreId: ExploreId): ThunkResult<void> {
       .then((response: PanelData) => {
         const { legacy, error, series } = response;
         if (error) {
-          dispatch(processQueryErrorsAction({ exploreId, response: error, datasourceId }));
+          dispatch(processErrorResults({ exploreId, response: error, datasourceId }));
           return;
         }
 
@@ -497,22 +516,219 @@ export function runQueries(exploreId: ExploreId): ThunkResult<void> {
         const nextHistory = updateHistory(history, datasourceId, queries);
         dispatch(historyUpdatedAction({ exploreId, history: nextHistory }));
         dispatch(
-          processQueryResultsAction({
+          processQueryResults({
             exploreId,
             latency,
             datasourceId,
             loadingState: LoadingState.Done,
             series: isLokiDataSource ? series : legacy,
-            delta: null,
           })
         );
-        dispatch(stateSaveAction());
+        dispatch(stateSave());
       })
       .catch(error => {
-        dispatch(processQueryErrorsAction({ exploreId, response: error, datasourceId }));
+        dispatch(processErrorResults({ exploreId, response: error, datasourceId }));
       });
   };
 }
+
+export const limitMessageRate = (
+  exploreId: ExploreId,
+  series: DataFrame[] | any[],
+  datasourceId: string
+): ThunkResult<void> => {
+  return (dispatch, getState) => {
+    dispatch(
+      processQueryResults({
+        exploreId,
+        latency: 0,
+        datasourceId,
+        loadingState: LoadingState.Streaming,
+        series,
+      })
+    );
+  };
+};
+
+export const processQueryResults = (config: {
+  exploreId: ExploreId;
+  latency: number;
+  datasourceId: string;
+  loadingState: LoadingState;
+  series?: DataQueryResponseData[];
+}): ThunkResult<void> => {
+  return (dispatch, getState) => {
+    const { exploreId, datasourceId, latency, loadingState, series } = config;
+    const { datasourceInstance, scanning, eventBridge } = getState().explore[exploreId];
+
+    // If datasource already changed, results do not matter
+    if (datasourceInstance.meta.id !== datasourceId) {
+      return;
+    }
+
+    const result = series || [];
+    const replacePreviousResults = loadingState === LoadingState.Done && series ? true : false;
+    const resultProcessor = new ResultProcessor(getState().explore[exploreId], replacePreviousResults, result);
+    const graphResult = resultProcessor.getGraphResult();
+    const tableResult = resultProcessor.getTableResult();
+    const logsResult = resultProcessor.getLogsResult();
+    const refIds = getRefIds(result);
+
+    // For Angular editors
+    eventBridge.emit('data-received', resultProcessor.getRawData());
+
+    // Clears any previous errors that now have a successful query, important so Angular editors are updated correctly
+    dispatch(resetQueryErrorAction({ exploreId, refIds }));
+
+    dispatch(
+      querySuccessAction({
+        exploreId,
+        latency,
+        loadingState,
+        graphResult,
+        tableResult,
+        logsResult,
+      })
+    );
+
+    // Keep scanning for results if this was the last scanning transaction
+    if (scanning) {
+      if (_.size(result) === 0) {
+        const range = getShiftedTimeRange(-1, getState().explore[exploreId].range);
+        dispatch(updateTime({ exploreId, absoluteRange: range }));
+        dispatch(runQueriesAction({ exploreId }));
+      } else {
+        // We can stop scanning if we have a result
+        dispatch(scanStopAction({ exploreId }));
+      }
+    }
+  };
+};
+
+export const processErrorResults = (config: {
+  exploreId: ExploreId;
+  response: any;
+  datasourceId: string;
+}): ThunkResult<void> => {
+  return (dispatch, getState) => {
+    const { exploreId, datasourceId } = config;
+    let { response } = config;
+    const { datasourceInstance, eventBridge } = getState().explore[exploreId];
+
+    if (datasourceInstance.meta.id !== datasourceId || response.cancelled) {
+      // Navigated away, queries did not matter
+      return;
+    }
+
+    // For Angular editors
+    eventBridge.emit('data-error', response);
+
+    console.error(response); // To help finding problems with query syntax
+
+    if (!instanceOfDataQueryError(response)) {
+      response = toDataQueryError(response);
+    }
+
+    dispatch(queryFailureAction({ exploreId, response }));
+  };
+};
+
+const toRawTimeRange = (range: TimeRange): RawTimeRange => {
+  let from = range.raw.from;
+  if (isDateTime(from)) {
+    from = from.valueOf().toString(10);
+  }
+
+  let to = range.raw.to;
+  if (isDateTime(to)) {
+    to = to.valueOf().toString(10);
+  }
+
+  return {
+    from,
+    to,
+  };
+};
+
+export const stateSave = (): ThunkResult<void> => {
+  return (dispatch, getState) => {
+    const { left, right, split } = getState().explore;
+    const orgId = getState().user.orgId.toString();
+    const replace = left && left.urlReplaced === false;
+    const urlStates: { [index: string]: string } = { orgId };
+    const leftUrlState: ExploreUrlState = {
+      datasource: left.datasourceInstance.name,
+      queries: left.queries.map(clearQueryKeys),
+      range: toRawTimeRange(left.range),
+      mode: left.mode,
+      ui: {
+        showingGraph: left.showingGraph,
+        showingLogs: true,
+        showingTable: left.showingTable,
+        dedupStrategy: left.dedupStrategy,
+      },
+    };
+    urlStates.left = serializeStateToUrlParam(leftUrlState, true);
+    if (split) {
+      const rightUrlState: ExploreUrlState = {
+        datasource: right.datasourceInstance.name,
+        queries: right.queries.map(clearQueryKeys),
+        range: toRawTimeRange(right.range),
+        mode: right.mode,
+        ui: {
+          showingGraph: right.showingGraph,
+          showingLogs: true,
+          showingTable: right.showingTable,
+          dedupStrategy: right.dedupStrategy,
+        },
+      };
+
+      urlStates.right = serializeStateToUrlParam(rightUrlState, true);
+    }
+
+    dispatch(updateLocation({ query: urlStates, replace }));
+    if (replace) {
+      dispatch(setUrlReplacedAction({ exploreId: ExploreId.left }));
+    }
+  };
+};
+
+export const updateTime = (config: {
+  exploreId: ExploreId;
+  rawRange?: RawTimeRange;
+  absoluteRange?: AbsoluteTimeRange;
+}): ThunkResult<void> => {
+  return (dispatch, getState) => {
+    const { exploreId, absoluteRange: absRange, rawRange: actionRange } = config;
+    const itemState = getState().explore[exploreId];
+    const timeZone = getTimeZone(getState().user);
+    const { range: rangeInState } = itemState;
+    let rawRange: RawTimeRange = rangeInState.raw;
+
+    if (absRange) {
+      rawRange = {
+        from: dateTimeForTimeZone(timeZone, absRange.from),
+        to: dateTimeForTimeZone(timeZone, absRange.to),
+      };
+    }
+
+    if (actionRange) {
+      rawRange = actionRange;
+    }
+
+    const range = getTimeRange(timeZone, rawRange);
+    const absoluteRange: AbsoluteTimeRange = { from: range.from.valueOf(), to: range.to.valueOf() };
+
+    getTimeSrv().init({
+      time: range.raw,
+      refresh: false,
+      getTimezone: () => timeZone,
+      timeRangeUpdated: (): any => undefined,
+    });
+
+    dispatch(changeRangeAction({ exploreId, range, absoluteRange }));
+  };
+};
 
 /**
  * Start a scan for more results using the given scanner.
@@ -526,7 +742,7 @@ export function scanStart(exploreId: ExploreId): ThunkResult<void> {
     // Scanning must trigger query run, and return the new range
     const range = getShiftedTimeRange(-1, getState().explore[exploreId].range);
     // Set the new range to be displayed
-    dispatch(updateTimeRangeAction({ exploreId, absoluteRange: range }));
+    dispatch(updateTime({ exploreId, absoluteRange: range }));
     dispatch(runQueriesAction({ exploreId }));
   };
 }
@@ -551,7 +767,7 @@ export function setQueries(exploreId: ExploreId, rawQueries: DataQuery[]): Thunk
 export function splitClose(itemId: ExploreId): ThunkResult<void> {
   return dispatch => {
     dispatch(splitCloseAction({ itemId }));
-    dispatch(stateSaveAction());
+    dispatch(stateSave());
   };
 }
 
@@ -575,7 +791,7 @@ export function splitOpen(): ThunkResult<void> {
       urlState,
     };
     dispatch(splitOpenAction({ itemState }));
-    dispatch(stateSaveAction());
+    dispatch(stateSave());
   };
 }
 
@@ -652,7 +868,7 @@ export function refreshExplore(exploreId: ExploreId): ThunkResult<void> {
     }
 
     if (update.range) {
-      dispatch(updateTimeRangeAction({ exploreId, rawRange: range.raw }));
+      dispatch(updateTime({ exploreId, rawRange: range.raw }));
     }
 
     // need to refresh ui state
