+++ public/app/features/explore/state/reducers.test.ts
@@ -26,12 +26,14 @@ import { serializeStateToUrlParam } from 'app/core/utils/explore';
 import TableModel from 'app/core/table_model';
 import { DataSourceApi, DataQuery } from '@grafana/ui';
 import { LogsModel, LogsDedupStrategy, LoadingState } from '@grafana/data';
+import { PanelQueryState } from '../../dashboard/state/PanelQueryState';
 
 describe('Explore item reducer', () => {
   describe('scanning', () => {
     it('should start scanning', () => {
       const initalState = {
         ...makeExploreItemState(),
+        queryState: null as PanelQueryState,
         scanning: false,
       };
 
@@ -40,12 +42,14 @@ describe('Explore item reducer', () => {
         .whenActionIsDispatched(scanStartAction({ exploreId: ExploreId.left }))
         .thenStateShouldEqual({
           ...makeExploreItemState(),
+          queryState: null as PanelQueryState,
           scanning: true,
         });
     });
     it('should stop scanning', () => {
       const initalState = {
         ...makeExploreItemState(),
+        queryState: null as PanelQueryState,
         scanning: true,
         scanRange: {},
       };
@@ -55,6 +59,7 @@ describe('Explore item reducer', () => {
         .whenActionIsDispatched(scanStopAction({ exploreId: ExploreId.left }))
         .thenStateShouldEqual({
           ...makeExploreItemState(),
+          queryState: null as PanelQueryState,
           scanning: false,
           scanRange: undefined,
         });
