+++ public/app/store/configureStore.ts
@@ -1,6 +1,5 @@
 import { createStore, applyMiddleware, compose, combineReducers } from 'redux';
 import thunk from 'redux-thunk';
-import { combineEpics, createEpicMiddleware } from 'redux-observable';
 import { createLogger } from 'redux-logger';
 import sharedReducers from 'app/core/reducers';
 import alertingReducers from 'app/features/alerting/state/reducers';
@@ -15,38 +14,8 @@ import usersReducers from 'app/features/users/state/reducers';
 import userReducers from 'app/features/profile/state/reducers';
 import organizationReducers from 'app/features/org/state/reducers';
 import { setStore } from './store';
-import { limitMessageRateEpic } from 'app/features/explore/state/epics/limitMessageRateEpic';
-import { stateSaveEpic } from 'app/features/explore/state/epics/stateSaveEpic';
-import { processQueryResultsEpic } from 'app/features/explore/state/epics/processQueryResultsEpic';
-import { processQueryErrorsEpic } from 'app/features/explore/state/epics/processQueryErrorsEpic';
-import {
-  DataSourceApi,
-  DataQueryResponse,
-  DataQuery,
-  DataSourceJsonData,
-  DataQueryRequest,
-  DataStreamObserver,
-} from '@grafana/ui';
-
-import {
-  TimeZone,
-  RawTimeRange,
-  TimeRange,
-  DateTimeInput,
-  FormatInput,
-  DateTime,
-  AbsoluteTimeRange,
-  dateTimeForTimeZone,
-} from '@grafana/data';
-import { Observable } from 'rxjs';
-import { getQueryResponse, getTimeRange } from 'app/core/utils/explore';
 import { StoreState } from 'app/types/store';
 import { toggleLogActionsMiddleware } from 'app/core/middlewares/application';
-import { timeEpic } from 'app/features/explore/state/epics/timeEpic';
-import { TimeSrv, getTimeSrv } from 'app/features/dashboard/services/TimeSrv';
-import { UserState } from 'app/types/user';
-import { getTimeZone } from 'app/features/profile/state/selectors';
-import { getShiftedTimeRange } from 'app/core/utils/timePicker';
 
 const rootReducers = {
   ...sharedReducers,
@@ -67,38 +36,6 @@ export function addRootReducer(reducers: any) {
   Object.assign(rootReducers, ...reducers);
 }
 
-export const rootEpic: any = combineEpics(
-  limitMessageRateEpic,
-  stateSaveEpic,
-  processQueryResultsEpic,
-  processQueryErrorsEpic,
-  timeEpic
-);
-
-export interface EpicDependencies {
-  getQueryResponse: (
-    datasourceInstance: DataSourceApi<DataQuery, DataSourceJsonData>,
-    options: DataQueryRequest<DataQuery>,
-    observer?: DataStreamObserver
-  ) => Observable<DataQueryResponse>;
-  getTimeSrv: () => TimeSrv;
-  getTimeRange: (timeZone: TimeZone, rawRange: RawTimeRange) => TimeRange;
-  getTimeZone: (state: UserState) => TimeZone;
-  getShiftedTimeRange: (direction: number, origRange: TimeRange, timeZone: TimeZone) => AbsoluteTimeRange;
-  dateTimeForTimeZone: (timezone?: TimeZone, input?: DateTimeInput, formatInput?: FormatInput) => DateTime;
-}
-
-const dependencies: EpicDependencies = {
-  getQueryResponse,
-  getTimeSrv,
-  getTimeRange,
-  getTimeZone,
-  getShiftedTimeRange,
-  dateTimeForTimeZone,
-};
-
-const epicMiddleware = createEpicMiddleware({ dependencies });
-
 export function configureStore() {
   const composeEnhancers = (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
   const rootReducer = combineReducers(rootReducers);
@@ -109,11 +46,10 @@ export function configureStore() {
   });
   const storeEnhancers =
     process.env.NODE_ENV !== 'production'
-      ? applyMiddleware(toggleLogActionsMiddleware, thunk, epicMiddleware, logger)
-      : applyMiddleware(thunk, epicMiddleware);
+      ? applyMiddleware(toggleLogActionsMiddleware, thunk, logger)
+      : applyMiddleware(thunk);
 
   const store = createStore(rootReducer, {}, composeEnhancers(storeEnhancers));
   setStore(store);
-  epicMiddleware.run(rootEpic);
   return store;
 }
